move error components to their own file
